separated position update to its own loop. + 20 fps...

+++ SDL_Playground/Player.cpp
@@ -5,13 +5,13 @@
 
 void Player::Update()
 {
-    float moveInputX = -input->A + input->D;
-    float moveInputY = -input->W + input->S;
+    float moveInputX = static_cast<float>(-input->A + input->D);
+    float moveInputY = static_cast<float>(-input->W + input->S);
 
     posX += moveInputX * speed * dTime;
     posY += moveInputY * speed * dTime;
         
     SDL_SetRenderDrawColor(renderer, 100, 100, 100, 255);
-    SDL_Rect playerRect = {(int)posX, (int)posY, 32, 32};
+    SDL_Rect playerRect = {static_cast<int>(posX), static_cast<int>(posY), 32, 32};
     SDL_RenderFillRect(renderer, &playerRect);
 }
